Add files via upload
The function has been significantly updated to streamline the path creation and boolean ops and is more flexible.  All of the path lines are now inflated using shapely, with a lot of boolean operation done using shapely polygons rather than the 3D meshes.  The function will now process inputs with no roads, waterways or waterbodies, and footpath can be input as a gpx file in which case no footpaths are read from OSM.  Mesh boolean ops are all performed in Blender now, and with the other changes the function usually runs in a few minutes, rather than hours.  The paths now include some smoothing and the buffer function is used to eliminate sections that are too small or thin to print, although there is still some room for improvement here.